Add support for setting the RAM requirements as specified by the user
This is important for environments where there are different slots sizes per host. Thus you can define the auto-slot split size to the largest value then rely on the RAM request to ensure you don't get allocated to small a node if your RAM requirements are >small node but <big node
The ram required must be divided up by the number of slots requested (when in a PE) as the RAM limit is applied to each slot.